feat(image-information): add a close icon leading back to the image-viewer
This component is loaded when you click on `More information` button from the `InformationPanel`.
It will be then handled by the router and displayed.
AS it is the only case, I assume it's ok to add a simple `router-link` pointing back to the ImageViewer for the same image.
If required, the current behavior can be changed to [use the previous route instead](https://stackoverflow.com/a/53789212).

https://gitlab.cytom.in/cm/rnd/issue_triage/-/issues/468