fix: set API keys in the store after update user
This is a workaround regarding the current behavior of API.
If we don't update the keys (in this case the "public" one)
in the store after apiEnabled is set to false, uploading the images
will not work anymore.

It's not great and the whole public/private keys concept should be
refactored to some extent on pims / core / web-ui.

cm/rnd/cytomine/frontend/web-ui-ee#22